Changed an unsigned int to a size_t for storing the size of a container.

+++ trunk/delta3d/src/dtGame/gamemanager.cpp
@@ -791,7 +791,7 @@ namespace dtGame
    void GameManager::GetAllComponents(std::vector<const GMComponent*>& toFill) const
    {
       toFill.clear();
-      unsigned int componentListSize = mComponentList.size();
+      size_t componentListSize = mComponentList.size();
       toFill.reserve(componentListSize);
 
       for (unsigned i = 0; i < componentListSize; ++i)
